Update step002_csiinblurout.py
corrected rescale width height data type

+++ python/step002_csiinblurout.py
@@ -28,7 +28,7 @@ while True:
     with vpi.Backend.CUDA:
         p1_vpiimg0 = vpi_img0.convert(vpi.Format.U8) \
                                 .box_filter(5, border=vpi.Border.ZERO) \
-                                .rescale((vpi_img0.width*0.5, vpi_img0.height*0.5), interp=vpi.Interp.LINEAR, border=vpi.Border.ZERO) \
+                                .rescale((int(vpi_img0.width*0.5), int(vpi_img0.height*0.5)), interp=vpi.Interp.LINEAR, border=vpi.Border.ZERO) \
                                 .box_filter(5, border=vpi.Border.ZERO)
         p1_vpiimg1 = vpi_img1.convert(vpi.Format.U8) \
                                 .box_filter(5, border=vpi.Border.ZERO) \
